Renamed planner wrappers
