+++ crp_APL/plan_motion_planning/include/plan_motion_planning/motionHandler.hpp
@@ -1,7 +1,6 @@
 #ifndef CRP_APL_MOTION_HANDLER_PLANMOTIONPLANNING_HPP
 #define CRP_APL_MOTION_HANDLER_PLANMOTIONPLANNING_HPP
 
-
 #include <rclcpp/rclcpp.hpp>
 #include <tier4_planning_msgs/msg/scenario.hpp>
 #include <autoware_planning_msgs/msg/trajectory.hpp>
