+++ crp_APL/planners/plan_lat_lane_follow/include/plan_lat_lane_follow/wrapperPlanLatLaneFollow.hpp
@@ -1,0 +1,22 @@
+#ifndef CRP_APL_WRAPPER_PLAN_LAT_LANE_FOLLOW_PLANLATLANEFOLLOW_HPP
+#define CRP_APL_WRAPPER_PLAN_LAT_LANE_FOLLOW_PLANLATLANEFOLLOW_HPP
+
+#include "planner_base/wrapperBase.hpp"
+
+namespace crp
+{
+namespace apl
+{
+
+class WrapperPlanLatLaneFollow : public WrapperBase
+{
+public:
+    WrapperPlanLatLaneFollow();
+
+private:
+    void plan(const PlannerInput & input, PlannerOutput & output) override;
+};
+
+} // namespace apl
+} // namespace crp
+#endif // CRP_APL_WRAPPER_PLAN_LAT_LANE_FOLLOW_PLANLATLANEFOLLOW_HPP
