+++ crp_APL/planners/plan_lon_emergency/CMakeLists.txt
@@ -13,7 +13,7 @@ find_package(planner_base REQUIRED)
 include_directories(include ${planner_base_INCLUDE_DIRS})
 
 
-add_executable(plan_lon_emergency src/wrapper.cpp)
+add_executable(plan_lon_emergency src/wrapperPlanLonEmergency.cpp)
 ament_target_dependencies(plan_lon_emergency rclcpp planner_base)
 
 install(TARGETS
