+++ crp_APL/planners/plan_lon_emergency/include/plan_lon_emergency/wrapperPlanLonEmergency.hpp
@@ -1,0 +1,22 @@
+#ifndef CRP_APL_WRAPPER_PLAN_LON_EMERGENCY_PLANLONEMERGENCY_HPP
+#define CRP_APL_WRAPPER_PLAN_LON_EMERGENCY_PLANLONEMERGENCY_HPP
+
+#include "planner_base/wrapperBase.hpp"
+
+namespace crp
+{
+namespace apl
+{
+
+class WrapperPlanLonEmergency : public WrapperBase
+{
+public:
+    WrapperPlanLonEmergency();
+
+private:
+    void plan(const PlannerInput & input, PlannerOutput & output) override;
+};
+
+} // namespace apl
+} // namespace crp
+#endif // CRP_APL_WRAPPER_PLAN_LON_EMERGENCY_PLANLONEMERGENCY_HPP
