--- crp_APL/planners/plan_lon_intelligent_speed_adjust/include/plan_lon_intelligent_speed_adjust/wrapper.hpp
@@ -1,22 +1,0 @@
-#ifndef CRP_APL_PLAN_LON_INTELLIGENT_SPEED_ADJUST_WRAPPER_PLANLATLANEFOLLOW_HPP
-#define CRP_APL_PLAN_LON_INTELLIGENT_SPEED_ADJUST_WRAPPER_PLANLATLANEFOLLOW_HPP
-
-#include "planner_base/wrapperBase.hpp"
-
-namespace crp
-{
-namespace apl
-{
-
-class PlanLonIntelligentSpeedAdjustWrapper : public WrapperBase
-{
-public:
-    PlanLonIntelligentSpeedAdjustWrapper();
-
-private:
-    void plan(const PlannerInput & input, PlannerOutput & output) override;
-};
-
-} // namespace apl
-} // namespace crp
-#endif // CRP_APL_PLAN_LON_INTELLIGENT_SPEED_ADJUST_WRAPPER_PLANLATLANEFOLLOW_HPP
