+++ crp_APL/planners/plan_lon_intelligent_speed_adjust/include/plan_lon_intelligent_speed_adjust/wrapperPlanLonIntelligentSpeedAdjust.hpp
@@ -1,0 +1,22 @@
+#ifndef CRP_APL_WRAPPER_PLAN_LON_INTELLIGENT_SPEED_ADJUST_PLANLATLANEFOLLOW_HPP
+#define CRP_APL_WRAPPER_PLAN_LON_INTELLIGENT_SPEED_ADJUST_PLANLATLANEFOLLOW_HPP
+
+#include "planner_base/wrapperBase.hpp"
+
+namespace crp
+{
+namespace apl
+{
+
+class WrapperPlanLonIntelligentSpeedAdjust : public WrapperBase
+{
+public:
+    WrapperPlanLonIntelligentSpeedAdjust();
+
+private:
+    void plan(const PlannerInput & input, PlannerOutput & output) override;
+};
+
+} // namespace apl
+} // namespace crp
+#endif // CRP_APL_WRAPPER_PLAN_LON_INTELLIGENT_SPEED_ADJUST_PLANLATLANEFOLLOW_HPP
